backend/requirements.txt
# =============================================================================
# DEPENDENCIES - Backend FastAPI (Phase 2+)
# =============================================================================

# --- Web Framework ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12      # Upload file support

# --- Database ---
sqlalchemy==2.0.35
alembic==1.14.0               # Database migrations (Phase 2+)
psycopg2-binary==2.9.10       # PostgreSQL adapter

# --- Settings & Validation ---
pydantic[email]==2.9.2         # [email] bao gồm email-validator cho EmailStr
pydantic-settings==2.6.1      # Load settings từ .env

# --- Authentication ---
python-jose[cryptography]==3.3.0  # JWT encode/decode (Mock Azure AD)
passlib[bcrypt]==1.7.4            # Password hashing utility

# --- Cache & Message Queue ---
redis==5.2.0
celery==5.4.0                 # Background task queue (Phase 3)

# --- AI / RAG (Phase 3) ---
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
pymupdf==1.24.9           # Extract text từ PDF (import as fitz)
python-docx==1.1.2        # Extract text từ DOCX

# --- HTTP Client ---
httpx==0.28.0                 # Gọi API của Ollama & LocalRecall

# --- Vector Database Client ---
# Phải match major version với image `chromadb/chroma` ở docker-compose.
# Server >= 1.x bỏ field "_type" trong CollectionConfiguration JSON → client 0.x
# parse fail với KeyError: '_type'. Giữ pin cùng major để tránh tái diễn.
chromadb>=1.0.0,<2.0.0

# --- Utilities ---
python-dateutil==2.9.0
